move changes for pgx support into new driver package
